+++ mayan/apps/acls/apps.py
@@ -2,7 +2,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.common.apps import MayanAppConfig
 from mayan.apps.common.classes import ModelCopy
-from mayan.apps.common.menus import menu_object, menu_secondary, menu_tools
+from mayan.apps.common.menus import menu_object, menu_secondary, menu_setup
 from mayan.apps.events.classes import EventModelRegistry, ModelEventType
 from mayan.apps.events.links import (
     link_events_for_object, link_object_event_types_user_subcriptions_list
@@ -11,7 +11,7 @@ from mayan.apps.navigation.classes import SourceColumn
 from mayan.apps.views.html_widgets import ObjectLinkWidget
 
 from .classes import ModelPermission
-from .events import event_acl_created, event_acl_edited
+from .events import event_acl_deleted, event_acl_edited
 from .links import (
     link_acl_create, link_acl_delete, link_acl_permissions,
     link_global_acl_list
@@ -30,7 +30,9 @@ class ACLsApp(MayanAppConfig):
         super().ready()
 
         AccessControlList = self.get_model(model_name='AccessControlList')
-        GlobalAccessControlListProxy = self.get_model(model_name='GlobalAccessControlListProxy')
+        GlobalAccessControlListProxy = self.get_model(
+            model_name='GlobalAccessControlListProxy'
+        )
 
         EventModelRegistry.register(model=AccessControlList)
 
@@ -41,7 +43,9 @@ class ACLsApp(MayanAppConfig):
         )
 
         ModelEventType.register(
-            event_types=(event_acl_created, event_acl_edited),
+            event_types=(
+                event_acl_deleted, event_acl_edited
+            ),
             model=AccessControlList
         )
 
@@ -76,6 +80,6 @@ class ACLsApp(MayanAppConfig):
         menu_secondary.bind_links(
             links=(link_acl_create,), sources=('acls:acl_list',)
         )
-        menu_tools.bind_links(
+        menu_setup.bind_links(
             links=(link_global_acl_list,)
         )