+++ mayan/apps/acls/events.py
@@ -9,6 +9,9 @@ namespace = EventTypeNamespace(
 event_acl_created = namespace.add_event_type(
     label=_('ACL created'), name='acl_created'
 )
+event_acl_deleted = namespace.add_event_type(
+    label=_('ACL deleted'), name='acl_deleted'
+)
 event_acl_edited = namespace.add_event_type(
     label=_('ACL edited'), name='acl_edited'
 )