+++ mayan/apps/acls/models.py
@@ -2,19 +2,24 @@ import logging
 
 from django.contrib.contenttypes.fields import GenericForeignKey
 from django.contrib.contenttypes.models import ContentType
-from django.db import models, transaction
+from django.db import models
 from django.urls import reverse
 from django.utils.translation import ugettext_lazy as _
 
+from mayan.apps.common.model_mixins import ExtraDataModelMixin
+from mayan.apps.events.classes import (
+    EventManagerMethodAfter, EventManagerSave
+)
+from mayan.apps.events.decorators import method_event
 from mayan.apps.permissions.models import Role, StoredPermission
 
-from .events import event_acl_created, event_acl_edited
+from .events import event_acl_created, event_acl_deleted, event_acl_edited
 from .managers import AccessControlListManager
 
 logger = logging.getLogger(name=__name__)
 
 
-class AccessControlList(models.Model):
+class AccessControlList(ExtraDataModelMixin, models.Model):
     """
     ACL means Access Control List it is a more fine-grained method of
     granting access to objects. In the case of ACLs, they grant access using
@@ -62,6 +67,14 @@ class AccessControlList(models.Model):
             'role': self.role,
         }
 
+    @method_event(
+        event_manager_class=EventManagerMethodAfter,
+        event=event_acl_deleted,
+        target='content_object'
+    )
+    def delete(self, *args, **kwargs):
+        return super().delete(*args, **kwargs)
+
     def get_absolute_url(self):
         return reverse(
             viewname='acls:acl_permissions', kwargs={'acl_id': self.pk}
@@ -73,33 +86,32 @@ class AccessControlList(models.Model):
         )
 
     def permissions_add(self, queryset, _user=None):
-        with transaction.atomic():
-            event_acl_edited.commit(
-                actor=_user, target=self
-            )
-            self.permissions.add(*queryset)
+        self.permissions.add(*queryset)
+        event_acl_edited.commit(
+            actor=_user, target=self
+        )
 
     def permissions_remove(self, queryset, _user=None):
-        with transaction.atomic():
-            event_acl_edited.commit(
-                actor=_user, target=self
-            )
-            self.permissions.remove(*queryset)
+        self.permissions.remove(*queryset)
+        event_acl_edited.commit(
+            actor=_user, target=self
+        )
 
+    @method_event(
+        event_manager_class=EventManagerSave,
+        created={
+            'action_object': 'content_object',
+            'event': event_acl_created,
+            'target': 'self',
+        },
+        edited={
+            'action_object': 'content_object',
+            'event': event_acl_edited,
+            'target': 'self',
+        }
+    )
     def save(self, *args, **kwargs):
-        _user = kwargs.pop('_user', None)
-
-        with transaction.atomic():
-            is_new = not self.pk
-            super().save(*args, **kwargs)
-            if is_new:
-                event_acl_created.commit(
-                    actor=_user, target=self
-                )
-            else:
-                event_acl_edited.commit(
-                    actor=_user, target=self
-                )
+        return super().save(*args, **kwargs)
 
 
 class GlobalAccessControlListProxy(AccessControlList):