+++ mayan/apps/acls/tests/mixins.py
@@ -1,4 +1,5 @@
 from django.core.exceptions import ImproperlyConfigured
+from django.db.models import Q
 
 from mayan.apps.testing.tests.mixins import TestModelTestCaseMixin
 from mayan.apps.permissions.tests.mixins import (
@@ -18,10 +19,21 @@ class ACLAPIViewTestMixin:
         if extra_data:
             data.update(extra_data)
 
-        return self.post(
+        pk_list = list(AccessControlList.objects.values_list('pk', flat=True))
+
+        response = self.post(
             viewname='rest_api:accesscontrollist-list',
             kwargs=self.test_object_view_kwargs, data=data
         )
+
+        try:
+            self.test_acl = AccessControlList.objects.get(
+                ~Q(pk__in=pk_list)
+            )
+        except AccessControlList.DoesNotExist:
+            self.test_acl = None
+
+        return response
 
     def _request_test_acl_delete_api_view(self):
         return self.delete(
@@ -84,7 +96,7 @@ class ACLAPIViewTestMixin:
             }
         )
 
-    def _request_test_acl_permission_list_api_post_view(self):
+    def _request_test_acl_permission_add_api_view(self):
         return self.post(
             viewname='rest_api:accesscontrollist-permission-list', kwargs={
                 'app_label': self.test_object_content_type.app_label,
@@ -92,6 +104,14 @@ class ACLAPIViewTestMixin:
                 'object_id': self.test_object.pk,
                 'acl_id': self.test_acl.pk
             }, data={'permission_pk': self.test_permission.pk}
+        )
+
+    def _request_test_class_permission_list_api_view(self):
+        return self.get(
+            viewname='rest_api:class-permission-list', kwargs={
+                'app_label': self.test_object_content_type.app_label,
+                'model_name': self.test_object_content_type.model,
+            }
         )
 
 
@@ -173,12 +193,23 @@ class AccessControlListViewTestMixin:
         )
 
     def _request_test_acl_create_post_view(self):
-        return self.post(
+        pk_list = list(AccessControlList.objects.values_list('pk', flat=True))
+
+        response = self.post(
             viewname='acls:acl_create',
             kwargs=self.test_object_view_kwargs, data={
                 'role': self.test_role.pk
             }
         )
+
+        try:
+            self.test_acl = AccessControlList.objects.get(
+                ~Q(pk__in=pk_list)
+            )
+        except AccessControlList.DoesNotExist:
+            self.test_acl = None
+
+        return response
 
     def _request_test_acl_delete_view(self):
         return self.post(
@@ -195,6 +226,27 @@ class AccessControlListViewTestMixin:
         return self.get(
             viewname='acls:acl_permissions', kwargs={
                 'acl_id': self.test_acl.pk
+            }
+
+        )
+
+    def _request_test_acl_permission_add_view(self):
+        return self.post(
+            viewname='acls:acl_permissions', kwargs={
+                'acl_id': self.test_acl.pk,
+            }, data={
+                'available-submit': 'true',
+                'available-selection': self.test_permission.stored_permission.pk
+            }
+        )
+
+    def _request_test_acl_permission_remove_view(self):
+        return self.post(
+            viewname='acls:acl_permissions', kwargs={
+                'acl_id': self.test_acl.pk,
+            }, data={
+                'added-submit': 'true',
+                'added-selection': self.test_permission.stored_permission.pk
             }
         )