+++ mayan/apps/acls/tests/test_api.py
@@ -3,6 +3,7 @@ from rest_framework import status
 from mayan.apps.rest_api.tests.base import BaseAPITestCase
 
 from ..classes import ModelPermission
+from ..events import event_acl_created, event_acl_deleted, event_acl_edited
 from ..models import AccessControlList
 from ..permissions import permission_acl_edit, permission_acl_view
 
@@ -15,54 +16,77 @@ class ACLAPIViewTestCase(ACLTestMixin, ACLAPIViewTestMixin, BaseAPITestCase):
     def test_acl_create_api_view_no_permission(self):
         acl_count = AccessControlList.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_acl_create_api_view()
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
         self.assertEqual(AccessControlList.objects.count(), acl_count)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_acl_create_api_view_with_access(self):
         self.grant_access(
             obj=self.test_object, permission=permission_acl_edit
         )
 
+        self._clear_events()
+
         response = self._request_test_acl_create_api_view()
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        pk = response.data['id']
 
-        test_object_acl = self.test_object.acls.get(pk=pk)
-        self.assertEqual(
-            test_object_acl.role, self.test_role
-        )
-        self.assertEqual(
-            test_object_acl.content_object, self.test_object
-        )
-        self.assertEqual(
-            test_object_acl.permissions.count(), 0
-        )
+        self.assertEqual(self.test_acl.role, self.test_role)
+        self.assertEqual(self.test_acl.content_object, self.test_object)
+        self.assertEqual(self.test_acl.permissions.count(), 0)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+        self.assertEqual(events[0].action_object, self.test_object)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_acl)
+        self.assertEqual(events[0].verb, event_acl_created.id)
 
     def test_acl_create_api_view_extra_data_with_access(self):
         self.grant_access(
             obj=self.test_object, permission=permission_acl_edit
         )
 
+        self._clear_events()
+
         response = self._request_test_acl_create_api_view(
             extra_data={'permissions_pk_list': permission_acl_view.pk}
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        pk = response.data['id']
 
-        test_object_acl = self.test_object.acls.get(pk=pk)
-
+        self.assertEqual(self.test_acl.content_object, self.test_object)
+        self.assertEqual(self.test_acl.role, self.test_role)
         self.assertEqual(
-            test_object_acl.content_object, self.test_object
-        )
-        self.assertEqual(
-            test_object_acl.role, self.test_role
-        )
-        self.assertEqual(
-            test_object_acl.permissions.first(),
+            self.test_acl.permissions.first(),
             permission_acl_view.stored_permission
         )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+        self.assertEqual(events[0].action_object, self.test_object)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_acl)
+        self.assertEqual(events[0].verb, event_acl_created.id)
+
+    def test_acl_delete_api_view_no_permission(self):
+        self._create_test_acl()
+
+        acl_count = AccessControlList.objects.count()
+
+        self._clear_events()
+
+        response = self._request_test_acl_delete_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        self.assertEqual(AccessControlList.objects.count(), acl_count)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_acl_delete_api_view_with_access(self):
         self._create_test_acl()
@@ -71,16 +95,30 @@ class ACLAPIViewTestCase(ACLTestMixin, ACLAPIViewTestMixin, BaseAPITestCase):
 
         acl_count = AccessControlList.objects.count()
 
+        self._clear_events()
+
         response = self._request_test_acl_delete_api_view()
         self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
 
         self.assertEqual(AccessControlList.objects.count(), acl_count - 1)
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_object)
+        self.assertEqual(events[0].verb, event_acl_deleted.id)
+
     def test_acl_detail_api_view_no_permission(self):
         self._create_test_acl()
 
+        self._clear_events()
+
         response = self._request_test_acl_detail_api_view()
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_acl_detail_api_view_with_access(self):
         self._create_test_acl()
@@ -88,6 +126,8 @@ class ACLAPIViewTestCase(ACLTestMixin, ACLAPIViewTestMixin, BaseAPITestCase):
         self.grant_access(
             obj=self.test_object, permission=permission_acl_view
         )
+
+        self._clear_events()
 
         response = self._request_test_acl_detail_api_view()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -99,11 +139,19 @@ class ACLAPIViewTestCase(ACLTestMixin, ACLAPIViewTestMixin, BaseAPITestCase):
             response.data['role']['label'], self.test_role.label
         )
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
     def test_acl_list_api_view_no_permission(self):
         self._create_test_acl()
 
+        self._clear_events()
+
         response = self._request_test_acl_list_api_view()
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_acl_list_api_view_with_access(self):
         self._create_test_acl()
@@ -111,6 +159,8 @@ class ACLAPIViewTestCase(ACLTestMixin, ACLAPIViewTestMixin, BaseAPITestCase):
         self.grant_access(
             obj=self.test_object, permission=permission_acl_view
         )
+
+        self._clear_events()
 
         response = self._request_test_acl_list_api_view()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -121,16 +171,64 @@ class ACLAPIViewTestCase(ACLTestMixin, ACLAPIViewTestMixin, BaseAPITestCase):
             response=response, text=self.test_acl.role.label
         )
 
-    def test_acl_permission_delete_api_view_with_access(self):
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+
+class ACLPermissionAPIViewTestCase(
+    ACLTestMixin, ACLAPIViewTestMixin, BaseAPITestCase
+):
+    auto_create_acl_test_object = True
+
+    def test_acl_permission_add_api_view_no_permission(self):
+        self._create_test_acl()
+
+        self._clear_events()
+
+        response = self._request_test_acl_permission_add_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        self.assertFalse(
+            self.test_permission.stored_permission in self.test_acl.permissions.all()
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_acl_permission_add_api_view_with_access(self):
+        self._create_test_acl()
+
+        self.grant_access(
+            obj=self.test_object, permission=permission_acl_edit
+        )
+
+        self._clear_events()
+
+        response = self._request_test_acl_permission_add_api_view()
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        self.assertTrue(
+            self.test_permission.stored_permission in self.test_acl.permissions.all()
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_acl)
+        self.assertEqual(events[0].verb, event_acl_edited.id)
+
+    def test_acl_permission_detail_api_view_no_permission(self):
         self._create_test_acl()
         self.test_acl.permissions.add(self.test_permission.stored_permission)
 
-        self.grant_access(obj=self.test_object, permission=permission_acl_edit)
+        self._clear_events()
 
-        response = self._request_test_acl_permission_delete_api_view()
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        response = self._request_test_acl_permission_detail_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
-        self.assertEqual(self.test_acl.permissions.count(), 0)
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
     def test_acl_permission_detail_api_view_with_access(self):
         self._create_test_acl()
@@ -140,12 +238,30 @@ class ACLAPIViewTestCase(ACLTestMixin, ACLAPIViewTestMixin, BaseAPITestCase):
             obj=self.test_object, permission=permission_acl_view
         )
 
+        self._clear_events()
+
         response = self._request_test_acl_permission_detail_api_view()
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(
             response.data['pk'], self.test_permission.pk
         )
 
-    def test_acl_permission_list_api_get_view_with_access(self):
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_acl_permission_list_api_view_no_permission(self):
+        self._create_test_acl()
+        self.test_acl.permissions.add(self.test_permission.stored_permission)
+
+        self._clear_events()
+
+        response = self._request_test_acl_permission_list_api_get_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_acl_permission_list_api_view_with_access(self):
         self._create_test_acl()
         self.test_acl.permissions.add(self.test_permission.stored_permission)
 
@@ -153,28 +269,64 @@ class ACLAPIViewTestCase(ACLTestMixin, ACLAPIViewTestMixin, BaseAPITestCase):
             obj=self.test_object, permission=permission_acl_view
         )
 
+        self._clear_events()
+
         response = self._request_test_acl_permission_list_api_get_view()
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(
             response.data['results'][0]['pk'],
             self.test_permission.pk
         )
 
-    def test_acl_permission_list_api_post_view_with_access(self):
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_acl_permission_remove_api_view_no_permission(self):
         self._create_test_acl()
+        self.test_acl.permissions.add(self.test_permission.stored_permission)
 
-        self.grant_access(
-            obj=self.test_object, permission=permission_acl_edit
+        acl_permission_count = self.test_acl.permissions.count()
+
+        self._clear_events()
+
+        response = self._request_test_acl_permission_delete_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        self.assertEqual(
+            self.test_acl.permissions.count(), acl_permission_count
         )
 
-        response = self._request_test_acl_permission_list_api_post_view()
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
 
-        self.assertTrue(
-            self.test_permission.stored_permission in self.test_acl.permissions.all()
+    def test_acl_permission_remove_api_view_with_access(self):
+        self._create_test_acl()
+        self.test_acl.permissions.add(self.test_permission.stored_permission)
+
+        self.grant_access(obj=self.test_object, permission=permission_acl_edit)
+
+        acl_permission_count = self.test_acl.permissions.count()
+
+        self._clear_events()
+
+        response = self._request_test_acl_permission_delete_api_view()
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+
+        self.assertEqual(
+            self.test_acl.permissions.count(), acl_permission_count - 1
         )
 
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self.test_acl)
+        self.assertEqual(events[0].verb, event_acl_edited.id)
 
-class ClassPermissionAPIViewTestCase(ACLTestMixin, BaseAPITestCase):
+
+class ClassPermissionAPIViewTestCase(
+    ACLTestMixin, ACLAPIViewTestMixin, BaseAPITestCase
+):
     auto_create_test_object = True
 
     def test_class_permission_list_api_view(self):
@@ -184,12 +336,7 @@ class ClassPermissionAPIViewTestCase(ACLTestMixin, BaseAPITestCase):
             )
         ]
 
-        response = self.get(
-            viewname='rest_api:class-permission-list', kwargs={
-                'app_label': self.test_object_content_type.app_label,
-                'model_name': self.test_object_content_type.model,
-            }
-        )
+        response = self._request_test_class_permission_list_api_view()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
 
         response_permissions = [